UI/Model Change for the filters in Competency/Skill/Practice

+++ HM.GM.BAL/Model/GMCalculationParams.cs
@@ -29,6 +29,6 @@ namespace HM.GM.BAL.Model
 
         public double TotalCost { get; set; }
 
-        public int NoOfMinds { get; set; }
+        public double NoOfMinds { get; set; }
     }
 }
